+++ src/hooks/useGlobalState.tsx
