+++ src/main.tsx
@@ -1,7 +1,7 @@
 import ReactDOM from "react-dom/client";
 import App from "./App.tsx";
 import "./index.css";
-import { StateProvider } from "./pages/useGlobalState.tsx";
+import { StateProvider } from "./hooks/useGlobalState.tsx";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <StateProvider>
